스토리 등록시 유효성 검증 실패시 메세지를 세분화하고, 첨부파일을 선택사항으로 수정합니다.

+++ app/Http/Controllers/API/StoryController.php
@@ -12,6 +12,7 @@ use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class StoryController extends Controller
 {
@@ -40,14 +41,21 @@ class StoryController extends Controller
             ]);
         }
 
-        $valid = validator($request->only('title', 'message', 'attachment'), [
+        $validator = Validator::make($request->all(), [
             'title' => 'required|string|max:255',
-            'message' => 'required'
+            'message' => 'required',
+            'attachment' => 'sometimes|max:4096'
+        ], [
+            'title.required' => '제목을 입력해 주세요',
+            'title.max' => '제목은 255자 이내로 입력해 주세요',
+            'message.required' => '스토리를 입력해 주세요',
+            'attachment.max' => '첨부파일은 4Mb 이하여야 합니다'
         ]);
-        if ($valid->fails()) {
+
+        if ($validator->fails()) {
             return response()->json([
                 'result' => 'fail',
-                'message' => $valid->errors()->all()
+                'message' => $validator->errors()->all()
             ], Response::HTTP_BAD_REQUEST);
         }
 
@@ -65,25 +73,27 @@ class StoryController extends Controller
 
             // 첨부파일 업로드
             $attachment_url = $request->file('attachment');
-            $file = $request->file('attachment')->getClientOriginalName();
-            $extension = pathinfo($file, PATHINFO_EXTENSION);
-            $lowerExtentsion = strtolower($extension);
-            $fileName = $story->id.".".$lowerExtentsion;
-            $attachmentPathFileName = $this->S3_PATH_STORY_ATTACHMENT.$fileName;
+            if ($attachment_url) {
+                $file = $request->file('attachment')->getClientOriginalName();
+                $extension = pathinfo($file, PATHINFO_EXTENSION);
+                $lowerExtentsion = strtolower($extension);
+                $fileName = $story->id.".".$lowerExtentsion;
+                $attachmentPathFileName = $this->S3_PATH_STORY_ATTACHMENT.$fileName;
 
-            $exists = Storage::disk('s3')->exists($attachmentPathFileName);
-            if ($exists) {
-                Storage::disk('s3')->delete($attachmentPathFileName);
+                $exists = Storage::disk('s3')->exists($attachmentPathFileName);
+                if ($exists) {
+                    Storage::disk('s3')->delete($attachmentPathFileName);
+                }
+                Storage::disk('s3')->put($attachmentPathFileName, file_get_contents($attachment_url));
+
+                $attachment = new Attachment();
+                $attachment->file_path = $this->S3_PATH_STORY_ATTACHMENT;
+                $attachment->file_name = $fileName;
+                $attachment->save();
+
+                $story->attachment_id = $attachment->id;
+                $story->save();
             }
-            Storage::disk('s3')->put($attachmentPathFileName, file_get_contents($attachment_url));
-
-            $attachment = new Attachment();
-            $attachment->file_path = $this->S3_PATH_STORY_ATTACHMENT;
-            $attachment->file_name = $fileName;
-            $attachment->save();
-
-            $story->attachment_id = $attachment->id;
-            $story->save();
 
             DB::commit();
         } catch (Exception $e) {
